moved main jsx etc into components
