+++ src/renderer/index.html
@@ -9,6 +9,6 @@
 
   <body style="background-color: black">
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/components/main.jsx"></script>
   </body>
 </html>
