+++ src/renderer/src/components/Game.jsx
@@ -1,0 +1,15 @@
+import { GameCommands } from './inputs/GameCommands'
+import { GameFunctions } from './funcs/GameFunctions'
+import { SaveIndicator } from './ui/indicators/SaveIndicator'
+
+const Game = () => {
+  return (
+    <>
+      <GameFunctions />
+      <GameCommands />
+      <SaveIndicator />
+    </>
+  )
+}
+
+export { Game }
